refactor(skiplist): move level number generators to separate classes
